Start reorganizing into separate libraries

+++ libs/shared/include/gw/Actions.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 namespace tr {
-enum class ActionType : uint32_t {
+enum class ActionType : uint8_t {
   MoveForward = 0,
   MoveBackward,
   StrafeLeft,
@@ -14,7 +14,7 @@ enum class ActionType : uint32_t {
   Cancel
 };
 
-enum class StateType : uint32_t {
+enum class StateType : uint8_t {
   Action = 0,
   State,
   Range
